requirements.txt
mcpo
python-dotenv
httpx
fastmcp
